+++ brakeout/GameEngine-v2/Wall.java
@@ -5,7 +5,7 @@ import java.awt.Rectangle;
 
 public class Wall extends Sprite{
 
-	public Wall(double x, double y, double width, double height) {
+	public Wall(int x, int y, int width, int height) {
 		super(x, y, width, height);
 		
 	}
